+++ examples/rpl_udp/main.c
@@ -25,7 +25,7 @@
 #include "shell.h"
 #include "shell_commands.h"
 #include "board_uart0.h"
-#include "destiny.h"
+#include "tl_socket.h"
 #include "kernel.h"
 
 #include "demo.h"
