+++ examples/rpl_udp/udp.c
@@ -27,7 +27,7 @@
 #include "kernel.h"
 #include "thread.h"
 
-#include "destiny/socket.h"
+#include "tl_socket/socket.h"
 
 #include "net_help.h"
 
@@ -57,7 +57,7 @@ void init_udp_server(void)
     char buffer_main[UDP_BUFFER_SIZE];
     int32_t recsize;
     uint32_t fromlen;
-    int sock = destiny_socket(PF_INET6, SOCK_DGRAM, IPPROTO_UDP);
+    int sock = tl_socket(PF_INET6, SOCK_DGRAM, IPPROTO_UDP);
 
     memset(&sa, 0, sizeof(sa));
 
@@ -66,13 +66,13 @@ void init_udp_server(void)
 
     fromlen = sizeof(sa);
 
-    if (-1 == destiny_socket_bind(sock, &sa, sizeof(sa))) {
+    if (-1 == tl_socket_bind(sock, &sa, sizeof(sa))) {
         printf("Error bind failed!\n");
-        destiny_socket_close(sock);
+        tl_socket_close(sock);
     }
 
     while (1) {
-        recsize = destiny_socket_recvfrom(sock, (void *)buffer_main, UDP_BUFFER_SIZE, 0,
+        recsize = tl_socket_recvfrom(sock, (void *)buffer_main, UDP_BUFFER_SIZE, 0,
                                           &sa, &fromlen);
 
         if (recsize < 0) {
@@ -82,7 +82,7 @@ void init_udp_server(void)
         printf("UDP packet received, payload: %s\n", buffer_main);
     }
 
-    destiny_socket_close(sock);
+    tl_socket_close(sock);
 }
 
 /* UDP send command */
@@ -105,7 +105,7 @@ void udp_send(int argc, char **argv)
     strncpy(text, argv[2], sizeof (text));
     text[sizeof (text) - 1] = 0;
 
-    sock = destiny_socket(PF_INET6, SOCK_DGRAM, IPPROTO_UDP);
+    sock = tl_socket(PF_INET6, SOCK_DGRAM, IPPROTO_UDP);
 
     if (-1 == sock) {
         printf("Error Creating Socket!");
@@ -120,7 +120,7 @@ void udp_send(int argc, char **argv)
     memcpy(&sa.sin6_addr, &ipaddr, 16);
     sa.sin6_port = HTONS(SERVER_PORT);
 
-    bytes_sent = destiny_socket_sendto(sock, (char *)text,
+    bytes_sent = tl_socket_sendto(sock, (char *)text,
                                        strlen(text) + 1, 0, &sa,
                                        sizeof sa);
 
@@ -133,5 +133,5 @@ void udp_send(int argc, char **argv)
                                             &ipaddr));
     }
 
-    destiny_socket_close(sock);
+    tl_socket_close(sock);
 }
