+++ sys/net/transport_layer/tcp/Makefile
@@ -1,3 +1,5 @@
 MODULE:=$(shell basename $(CURDIR))
 
+INCLUDES += -I $(RIOTBASE)/sys/net/transport_layer/tl_socket
+
 include $(RIOTBASE)/Makefile.base
