+++ sys/net/transport_layer/tl_socket/Makefile
@@ -1,0 +1,7 @@
+MODULE:=$(shell basename $(CURDIR))
+
+INCLUDES += -I $(RIOTBASE)/sys/net/transport_layer/tl_socket
+INCLUDES += -I $(RIOTBASE)/sys/net/transport_layer/udp
+INCLUDES += -I $(RIOTBASE)/sys/net/transport_layer/tcp
+
+include $(RIOTBASE)/Makefile.base
