+++ sys/net/transport_layer/udp/Makefile
@@ -1,0 +1,5 @@
+MODULE:=$(shell basename $(CURDIR))
+
+INCLUDES += -I $(RIOTBASE)/sys/net/transport_layer/tl_socket
+
+include $(RIOTBASE)/Makefile.base
