+++ sys/posix/pnet/include/netinet/in.h
@@ -28,7 +28,7 @@
 #include <sys/socket.h>
 #include <arpa/inet.h>
 #include "ipv6.h"
-#include "destiny/socket.h"
+#include "tl_socket/socket.h"
 
 /**
  * IPv4 socket address type.
