+++ sys/posix/pnet/sys_socket.c
@@ -14,24 +14,24 @@
  */
 #include <errno.h>
 
-#include "destiny/socket.h"
+#include "tl_socket/socket.h"
 #include "fd.h"
 
 #include "sys/socket.h"
 
 int flagless_send(int fd, const void *buf, size_t len)
 {
-    return (int)destiny_socket_send(fd, buf, (uint32_t)len, 0);
+    return (int)tl_socket_send(fd, buf, (uint32_t)len, 0);
 }
 
 int flagless_recv(int fd, void *buf, size_t len)
 {
-    return (int)destiny_socket_recv(fd, buf, (uint32_t)len, 0);
+    return (int)tl_socket_recv(fd, buf, (uint32_t)len, 0);
 }
 
 int socket(int domain, int type, int protocol)
 {
-    int internal_socket = destiny_socket(domain, type, protocol);
+    int internal_socket = tl_socket(domain, type, protocol);
 
     if (internal_socket < 0) {
         errno = ENFILE;
@@ -39,7 +39,7 @@ int socket(int domain, int type, int protocol)
     }
 
     return fd_new(internal_socket, flagless_recv, flagless_send,
-                  destiny_socket_close);
+                  tl_socket_close);
 }
 
 
@@ -51,12 +51,12 @@ int socket(int domain, int type, int protocol)
 int accept(int socket, struct sockaddr *restrict address,
            socklen_t *restrict address_len)
 {
-    int res = sock_func_wrapper(destiny_socket_accept, socket,
+    int res = sock_func_wrapper(tl_socket_accept, socket,
                                 (sockaddr6_t *)address,
                                 (socklen_t *)address_len);
 
     if (res < 0) {
-        // destiny needs more granular error handling
+        // tl_socket needs more granular error handling
         errno = EOPNOTSUPP;
         return -1;
     }
@@ -66,11 +66,11 @@ int accept(int socket, struct sockaddr *restrict address,
 
 int bind(int socket, const struct sockaddr *address, socklen_t address_len)
 {
-    int res = sock_func_wrapper(destiny_socket_bind, socket,
+    int res = sock_func_wrapper(tl_socket_bind, socket,
                                 (sockaddr6_t *)address, address_len);
 
     if (res < 0) {
-        // destiny needs more granular error handling
+        // tl_socket needs more granular error handling
         errno = EOPNOTSUPP;
         return -1;
     }
@@ -80,11 +80,11 @@ int bind(int socket, const struct sockaddr *address, socklen_t address_len)
 
 int connect(int socket, const struct sockaddr *address, socklen_t address_len)
 {
-    int res = sock_func_wrapper(destiny_socket_connect, socket,
+    int res = sock_func_wrapper(tl_socket_connect, socket,
                                 (sockaddr6_t *)address, address_len);
 
     if (res < 0) {
-        // destiny needs more granular error handling
+        // tl_socket needs more granular error handling
         errno = ECONNREFUSED;
         return -1;
     }
@@ -101,10 +101,10 @@ int getsockopt(int socket, int level, int option_name,
 
 int listen(int socket, int backlog)
 {
-    int res = sock_func_wrapper(destiny_socket_listen, socket, backlog);
+    int res = sock_func_wrapper(tl_socket_listen, socket, backlog);
 
     if (res < 0) {
-        // destiny needs more granular error handling
+        // tl_socket needs more granular error handling
         errno = ECONNREFUSED;
         return -1;
     }
@@ -114,11 +114,11 @@ int listen(int socket, int backlog)
 
 ssize_t recv(int socket, void *buffer, size_t length, int flags)
 {
-    int32_t res = sock_func_wrapper(destiny_socket_recv, socket, buffer,
+    int32_t res = sock_func_wrapper(tl_socket_recv, socket, buffer,
                                     (uint32_t) length, flags);
 
     if (res < 0) {
-        // destiny needs more granular error handling
+        // tl_socket needs more granular error handling
         errno = ENOTCONN;
         return -1;
     }
@@ -130,13 +130,13 @@ ssize_t recvfrom(int socket, void *restrict buffer, size_t length, int flags,
                  struct sockaddr *restrict address,
                  socklen_t *restrict address_len)
 {
-    int32_t res = sock_func_wrapper(destiny_socket_recvfrom, socket, buffer,
+    int32_t res = sock_func_wrapper(tl_socket_recvfrom, socket, buffer,
                                     (uint32_t) length, flags,
                                     (sockaddr6_t *)address,
                                     (socklen_t *)address_len);
 
     if (res < 0) {
-        // destiny needs more granular error handling
+        // tl_socket needs more granular error handling
         errno = ENOTCONN;
         return -1;
     }
@@ -146,11 +146,11 @@ ssize_t recvfrom(int socket, void *restrict buffer, size_t length, int flags,
 
 ssize_t send(int socket, const void *buffer, size_t length, int flags)
 {
-    int32_t res = sock_func_wrapper(destiny_socket_send, socket, buffer,
+    int32_t res = sock_func_wrapper(tl_socket_send, socket, buffer,
                                     (uint32_t) length, flags);
 
     if (res < 0) {
-        // destiny needs more granular error handling
+        // tl_socket needs more granular error handling
         errno = ENOTCONN;
         return -1;
     }
@@ -161,13 +161,13 @@ ssize_t send(int socket, const void *buffer, size_t length, int flags)
 ssize_t sendto(int socket, const void *message, size_t length, int flags,
                const struct sockaddr *dest_addr, socklen_t dest_len)
 {
-    int32_t res = sock_func_wrapper(destiny_socket_sendto, socket, message,
+    int32_t res = sock_func_wrapper(tl_socket_sendto, socket, message,
                                     (uint32_t) length, flags,
                                     (sockaddr6_t *)dest_addr,
                                     (socklen_t)dest_len);
 
     if (res < 0) {
-        // destiny needs more granular error handling
+        // tl_socket needs more granular error handling
         errno = ENOTCONN;
         return -1;
     }
